Added a CNI test for networking statistics.
This is a veth CNI plugin that is written in bash. It creates a veth
virtual network pair, one end of the pair will be moved to container
network namespace.

The veth CNI plugin uses 203.0.113.0/24 subnet, it is reserved for
documentation and examples [rfc5737]. The plugin can allocate up to
128 veth pairs.

Review: https://reviews.apache.org/r/68355